+++ lib/flutter_map_supercluster.dart
@@ -6,6 +6,7 @@ export 'package:supercluster/supercluster.dart';
 export 'src/controller/marker_event.dart';
 export 'src/controller/supercluster_controller.dart';
 export 'src/controller/supercluster_state.dart';
+export 'src/layer/center_zoom_animation.dart';
 export 'src/layer/cluster_data.dart';
 export 'src/layer/supercluster_config.dart';
 export 'src/layer/supercluster_layer.dart';