+++ lib/src/layer/center_zoom_animation.dart
@@ -1,0 +1,7 @@
+class CenterZoomAnimation {
+  static const started = 'CenterZoomAnimation.started';
+  static const inProgress = 'CenterZoomAnimation.inProgress';
+  static const finished = 'CenterZoomAnimation.finished';
+
+  const CenterZoomAnimation._();
+}